+++ productdevelopment/src/main/java/com/htc/productdevelopment/repository/DepartmentRepository.java
@@ -1,8 +1,7 @@
-	package com.htc.productdevelopment.repository;
-	
-	import com.htc.productdevelopment.model.Department;
-	import org.springframework.data.jpa.repository.JpaRepository;
-	
-	public interface DepartmentRepository extends JpaRepository<Department, Integer> {
-	
-	}
+package com.htc.productdevelopment.repository;
+
+import com.htc.productdevelopment.model.Department;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
+}
